Fixup: Properly handle inserting to the rope
There were a couple of bugs with our implementations:
1. We were creating some unnecessary nodes in the
case of an insert that turns out to be an append
operation in the tree, without requiring some
node split.
2. The prefix count was not correctly updated.

While I am not convinced, we have a good algorithm
to propagate 2, it seems to work on some small use case.