Add a stack-based solution
This uses a custom stack to handle. While we could use a normal stack,
we would need to inspect it quite a lot to find out the previous
repeat ancestor. This ends up being easier to just handle this
with a custom DS (which also allows us to save those repeat
ancestors so we don't need to walk back).